Alterado para não limpar a collection de tools após passar pelo teste de tools --

+++ src/__tests__/utils/truncate.js
@@ -1,7 +1,10 @@
 const User = require("../../models/User");
 const Tool = require("../../models/Tool");
 
-module.exports = async () => {
+module.exports = async (clearToolsCollection = false) => {
   await User.deleteMany();
-  await Tool.deleteMany();
+
+  if (clearToolsCollection) {
+    await Tool.deleteMany();
+  }
 };
